Move errors outside of api directory.
